+++ test/fixtures/jvm/Main.kt
@@ -1,0 +1,10 @@
+package jvm
+
+import jvm.kotlin.a.Class
+
+class Main {
+    fun method() {
+        Class().method()
+        jvm.kotlin.b.Class().method()
+    }
+}
