+++ test/fixtures/jvm/java/Class.java
@@ -1,0 +1,9 @@
+package jvm.java;
+
+import jvm.kotlin.a.Class;
+
+public class Class {
+    public void method() {
+        new Class().method();
+    }
+}
