+++ test/fixtures/jvm/kotlin/a/Class.kt
@@ -1,4 +1,4 @@
-package kotlin.b
+package jvm.kotlin.a
 
 class Class {
     fun method() {
