--- test/fixtures/kotlin/Main.kt
@@ -1,10 +1,0 @@
-package kotlin
-
-import kotlin.a.Class
-
-class Main {
-    fun method() {
-        Class().method()
-        kotlin.b.Class().method()
-    }
-}
